+++ src/domain/book/entity/CleanCode.java
@@ -6,7 +6,7 @@ import domain.book.interfaces.ShippableBook;
 public class CleanCode extends Book implements PaperBook,ShippableBook {
     double weight;
     int stock;
-    public CleanCode(String title, String author, String publisher, String isbn, int yearOfPublication, int price, double weight,int stock) {
+    public CleanCode(String title, String author, String publisher, String isbn, int yearOfPublication, double price, double weight,int stock) {
         super(title, author, publisher, isbn, yearOfPublication, price);
         this.weight = weight;
         this.stock = stock;
